Refactor : Move PlayerGrowth to Ship
